% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generaFunzione.R
\name{chooseF}
\alias{chooseF}
\title{Genera una funzione random}
\usage{
chooseF(tipo, a = 9, b = 2)
}
\arguments{
\item{tipo}{di funzione}

\item{a}{il range}

\item{b}{quanti}
}
\value{
la funzione generata (deve diventare un oggetto di classe F)
}
\description{
Genera una funzione random
}
\examples{
chooseF("sin",3,2)
}

