% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funzioneCasuale.R
\name{funzioneCasuale}
\alias{funzioneCasuale}
\title{Genera una funzione random (oggetto di classe Funzione).}
\usage{
funzioneCasuale(tipo, a = 9, b = 2, finito = F, bignumber = 10^8,
  smallnumber = 10^-4)
}
\arguments{
\item{tipo}{di funzione}

\item{a}{il range dei parametri}

\item{b}{il numero di parametri}
}
\value{
la funzione generata (un oggetto di classe Funzione)
}
\description{
Per le funzioni periodiche il dominio viene ristretto ad un periodo,
per le altre funzioni si considera un sottointervallo del dominio in cui la
funzione assuma valori assoluti minori bignumber e  maggiori di smallnumber.
}
\examples{
funzioneCasuale("sin",3,2)
}

