% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generaHyp.R
\name{generaHyp}
\alias{generaHyp}
\title{Genera una funzione seno random (oggetto di classe Funzione).}
\usage{
generaHyp(a = 9, b = 2, finito = FALSE, xmax = 10)
}
\arguments{
\item{a}{il range dei parametri}

\item{b}{il numero di parametri}

\item{finito}{variabile logica per il dominio}
}
\value{
la funzione generata (un oggetto di classe Funzione)
}
\description{
Genera una funzione seno random (oggetto di classe Funzione).
}
\examples{
generasin(finito=TRUE)
}

