% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\docType{methods}
\name{initialize}
\alias{initialize}
\title{Constructor della Class Funzione.}
\usage{
\S4method{initialize}{Funzione}(.Object, fR, fTex, dominio)
}
\description{
Constructor della Class Funzione.
}
\section{Slots}{

\describe{
\item{\code{fR}}{an R formula}

\item{\code{fTex}}{a Tex formula}

\item{\code{dominio}}{the Domain}
}}

