% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuovaFunzione.R
\name{nuovaFunzione}
\alias{nuovaFunzione}
\title{Crea una nuova funzione}
\usage{
nuovaFunzione(fR, fTex, dominio)
}
\arguments{
\item{fR}{an R formula}

\item{fTex}{a Tex formula}

\item{dominio}{the Domain}
}
\value{
Una nuova Funzione
}
\description{
Crea una nuova funzione
}
\examples{
nuovaFunzione(f<-function(x) x^2,"x^2",c(-Inf,Inf))
}

