% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prodotto.R, R/prodottoClass.R, R/prodottocopy.R
\name{prodottoR}
\alias{prodottoR}
\title{Prodotto in R
#' @author Maurizio Rinaldi}
\usage{
prodottoR(A, B)

prodottoR(A, B)

prodottoR(A, B)
}
\arguments{
\item{A}{una funzione di una variabile in  R}

\item{B}{una funzione di una variabile in  R}

\item{A}{un oggetto di classe funzione}

\item{B}{un oggetto di classe funzione}

\item{A}{un oggetto di classe funzione}

\item{B}{un oggetto di classe funzione}
}
\value{
La funzione prodotto

La funzione prodotto

La funzione prodotto
}
\description{
Prodotto in R
#' @author Maurizio Rinaldi

Prodotto in R

Prodotto in R
}

