% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prodottoTex.R
\name{prodottoTex}
\alias{prodottoTex}
\title{tex prodotto}
\usage{
prodottoTex(x1, x2)
}
\arguments{
\item{x1}{A Funzione Tex}

\item{x2}{B Funzione Tex}
}
\value{
The product of the two functions
}
\description{
tex prodotto
}
\examples{
x1="e^x"
x2="log(x)"
prodottoTex(x1,x2)
}

