% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sceltarange.R
\name{sceltarange}
\alias{sceltarange}
\title{Definizione di sceltarange}
\usage{
sceltarange(m = 1, da, a, escluso = c(), logica = T)
}
\arguments{
\item{m}{quanti valori}

\item{da}{valore minimo dei parametri}

\item{a}{valore massimo dei parametri}

\item{escluso}{i valori da eliminares}

\item{logica}{se il campione prevede ripetizioni}
}
\value{
m numeri nell'intervallo specificato
}
\description{
Definizione di sceltarange
}

