% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{studentibiotec}
\alias{studentibiotec}
\title{Studenti di Biotecnologie}
\format{A data frame with 41 rows and 8 variabili:
\describe{
  \item{SESSO}{M/F}
  \item{PESO}{peso in kg}
  \item{ALTEZZA}{Altezza in metri}
  \item{OCCHI}{Colore degli occhi}
  \item{CAPELLI}{Colore dei capelli}
  \item{SCARPE}{numero di scarpe}
  \item{hMADRE}{Altezza della madre}
  \item{hPADRE}{Altezza del padre}
  
}}
\usage{
studentibiotec
}
\description{
#' Un dataset che contiene informazioni su 41 studenti di Biotec
}
\keyword{datasets}

