% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{studentifarmacia}
\alias{studentifarmacia}
\title{Studenti di Farmacia}
\format{Un data frame con 41 righe e 8 variabili:
\describe{
  \item{SEX}{M/F}
  \item{W}{peso in kg}
  \item{H}{Altezza in metri}
  \item{Eyes}{Colore degli occhi}
  \item{Hair}{Colore dei capelli}
  \item{Sh}{numero di scarpe}
  \item{hM}{Altezza della madre}
  \item{hF}{Altezza del padre}
}}
\usage{
studentifarmacia
}
\description{
I dati: Un dataset che contiene informazioni su 55 studenti di Farmacia
}
\keyword{datasets}

