% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniforme.R
\name{uniforme}
\alias{uniforme}
\title{Definizione della classe uniforme}
\usage{
uniforme(n, xmin = 0, xmax = 1000, step = 10)
}
\arguments{
\item{n}{quanti valori}

\item{xmin}{minimo valore dell'estremo sinistro}

\item{xmax}{massimo valore dell'estremo sinistro}

\item{step}{incremento}
}
\value{
n punti con distribuzione uniforme
}
\description{
Definizione della classe uniforme
}
\author{
Maurizio Rinaldi
}

